\name{air}
\docType{data}
\alias{air}
\title{Air quality measurements in an underground train station}

\description{
This data set contains the hourly mean air
quality measurements from 01 January 2013 to 31 December 2016 in the
Chatelet underground train station in the Paris metro. 
}

\usage{data(air)}

\format{A matrix with 35039 rows and 8 columns.
Each row corresponds to an hourly measurement. 
The first column is the date (yyyy-mm-dd),
the second is the time (hh:mm),
the third is the nitric oxide NO concentration (g/m3),
the fourth is the nitrogen dioxide NO\eqn{_2}{2} concentration (g/m3),
the fifth is the concentration of particulate matter less than 10
microns PM10 (ppm),
the sixth is the carbon dioxide concentration CO\eqn{_2}{2} (g/m3),
the seventh is the temperature (degrees Celsius),
the eighth is the relative humidity (percentage). }

\source{
RATP (2016) Qualite de l'air mesuree dans la station Chatelet, 
\url{https://data.iledefrance.fr/explore/dataset/qualite-de-lair-mesuree-dans-la-station-chatelet}. Regie autonome des transports parisiens - Departement Developpement, Innovation et Territoires. Accessed 2017-09-27.
}

\keyword{datasets}

