\name{cardio}
\docType{data}
\alias{cardio}
\title{Foetal cardiotocograms}

\description{
  This data set contains the cardiotocographic measurements from healthy,
  suspect and pathological foetuses. 
}

\usage{data(cardio)}

\format{A matrix with 2126 rows and 8 columns.
  Each row corresponds to a foetal cardiotocogram. The
  class label for the foetal state is the last column: N = normal, S =
  suspect, P = pathological. Details for all variables are found in the
  link below.  
}

\source{
Lichman, M. (2013) UCI Machine learning repository: cardiotocography data set.
%\url{http://archive.ics.uci.edu/ml/datasets/Cardiotocography}.
University of California, Irvine, School of Information and Computer
Sciences. Accessed 2017-05-18.
}

\keyword{datasets}

