\name{tempb}
\docType{data}
\alias{tempb}
\title{Daily temperature}

\description{
  This data set contains the daily minimum and maximum temperatures from
  the weather station in Badajoz, Spain, from 1 January 1955 to 31 December
  2015.  
}

\usage{data(tempb)}

\format{A matrix with  21908 rows and 5 columns.
  Each row corresponds to a daily measurement. 
  The first column is the year (yyyy), 
  the second is the month (mm),
  the third is the day (dd), 
  the fourth is the minimum temperature (degrees Celsius),
  the fifth is the maximum temperature (degrees Celsius).
}

\source{
Menne, M. J., Durre, I., Vose, R. S., Gleason, B. E. & Houston,
T. (2012) An overview of the global historical climatology network-daily
database, \emph{Journal of Atmospheric and Oceanic Technology}
\bold{429}, 897 - 910. \url{https://climexp.knmi.nl/selectdailyseries.cgi}. Accessed 2016-10-20.
}

\keyword{datasets}

