\name{unicef}
\docType{data}
\alias{unicef}
\title{Unicef child mortality - life expectancy data}

\description{
This data set contains the number of deaths of children under 5 years of
age per 1000 live births and the average life expectancy (in years) at birth
for 73 countries with GNI (Gross National Income) less
than 1000 US dollars per annum per capita.
}

\usage{data(unicef)}

\format{A matrix with 2 columns and 73 rows. Each row corresponds to a country.
The first column is the under 5 mortality rate and the second is the average
life expectancy.}

\source{
Unicef (2003). \emph{State of the World's Children Report 2003}, Oxford University Press, for Unicef.
}

\keyword{datasets}

