\name{worldbank}

\docType{data}

\alias{worldbank}

\title{Development indicators from the World Bank Group}

\description{
This data set contains six development indicators for \eqn{n=218}{n=218} national
entities for the year 2011, which is the latest year for which they are
consistently available. These indicators are the carbon dioxide 
emissions per capita (thousands Kg), the gross domestic product GDP per
capita (thousands of current USD), the annual GDP growth rate (\%), the
annual inflation rate (\%), the number of internet users in the
population (\%) and the added value of agricultural production as a
ratio of the total GDP (\%). 
}

\usage{data(worldbank)}

\format{A matrix with 7 columns and 218 rows. Each row corresponds to a country.
The first column is the country, the second is the per capita carbon dioxide
emissions, the third is the per capita GDP, the fourth is the GDP growth
rate, the fifth is the inflation rate, the sixth is the percentage of
internet users and the seventh is the added value agricultural
production. }

\source{
World Bank Group (2016) World development indicators. 
\url{http://databank.worldbank.org/data/reports.aspx?
source=world-development-indicators}. Accessed 2016-10-03.
}


\keyword{datasets}

