\name{grevillea}

\docType{data}

\alias{grevillea}

\title{Geographical locations of grevillea plants}

\description{
This data set contains the geographical locations 
of the specimens of \emph{Grevillea uncinulata},
more commonly known as the Hook leaf grevillea, which is an endemic floral
species to south Western Australia. This region is
one of the 25 `biodiversity hotspots' which are `areas featuring exceptional
concentrations of endemic species and experiencing exceptional loss of habitat'.
}

\usage{data(grevillea)}

\format{A matrix with 222 rows and 2 columns. Each row corresponds to a
  observed plant. The first column is the longitude (decimal degrees),
  the second is the latitude (decimal degrees). }

\source{
CSIRO (2016) Atlas of Living Australia: \emph{Grevillea uncinulata Diels},
\url{http://bie.ala.org.au/species/urn:lsid:biodiversity.org.au:apni.taxon:249319}. Commonwealth Scientific and Industrial Research Organisation. Accessed 2016-03-11. 
}


\keyword{datasets}

