\name{ks-internal}

\alias{Sdr}
\alias{dfltCounts}
\alias{drvkde}
\alias{kde.boundary}
\alias{Sdrv}
\alias{mur}
\alias{nur}
\alias{nurs}
\alias{Qr}
\alias{matrix.sqrt}

\title{Internal functions in the ks library}
\description{
These functions are user-level but which the user is not required to use directly.
}


\keyword{internal}
