\name{ks-internal}
\alias{Hlscv.ab}
\alias{Hlscv.pc} 
\alias{Hlscv.vkde} 
\alias{Psi4.elem} 
\alias{Psi4.list} 
\alias{RKfun} 
\alias{Theta6.elem}
\alias{bcv.mat}
\alias{clust.tree} 
\alias{dmvnorm.2d.sum}
\alias{dmvnorm.2d.sum.pc}
\alias{dmvnorm.4d.sum}
\alias{dmvnorm.4d.sum.old} 
\alias{dmvnorm.4d.sum.perm} 
\alias{dmvnorm.6d.sum}
\alias{dmvnorm.deriv.2d} 
\alias{dmvnorm.deriv.2d.sum} 
\alias{dmvnorm.deriv.2d.xxt.sum}
\alias{dmvnorm.deriv.4d} 
\alias{dmvnorm.deriv.4d.sum} 
\alias{dmvnorm.deriv.6d}
\alias{dmvnorm.deriv.6d.sum}
\alias{dmvnorm.deriv.diag.old.4d} 
\alias{dmvt} 
\alias{dupl} 
\alias{elem}
\alias{find.gridpts} 
\alias{gamse.even.2d} 
\alias{gamse.odd.2d} 
\alias{gamse.scv.2d} 
\alias{gamse.scv.4d}
\alias{gamse.scv.6d} 
\alias{gsamse.2d} 
\alias{gsamse.4d} 
\alias{gsamse.6d} 
\alias{invdupl} 
\alias{invvec} 
\alias{invvech}
\alias{invvech.cat} 
\alias{is.even}
\alias{ise.mixt.pc}
\alias{kde.grid.2d} 
\alias{kde.pc.grid.2d}
\alias{kde.pc.points} 
\alias{kde.points}
\alias{lambda}
\alias{list.length} 
\alias{lscv.mat} 
\alias{lscv.mat.ab}
\alias{lscv.mat.pc} 
\alias{make.grid} 
\alias{make.supp} 
\alias{matrix.sqrt} 
\alias{mise.mixt.pc}
\alias{moments.mixt} 
\alias{mvpeaks.2d}
\alias{nearest.mode} 
\alias{omega}
\alias{optnum.clust} 
\alias{permut}
\alias{permute}
\alias{plot.pvkde} 
\alias{plotkde.2d}
\alias{plotmixt}
\alias{pre.abram}
\alias{pre.clust} 
\alias{pre.sain}
\alias{pre.scale.pc} 
\alias{pre.sphere.pc} 
\alias{pre.vkde}
\alias{psifun1.2d} 
\alias{psifun1.4d} 
\alias{psifun1.6d}
\alias{psifun2.2d} 
\alias{psifun2.4d} 
\alias{psifun2.6d} 
\alias{psifun2.diag.6d} 
\alias{psimat.2d} 
\alias{psimat.4d}
\alias{psimat.6d}
\alias{psimat.diag.6d} 
\alias{psins.2d} 
\alias{psins.4d}
\alias{psins.4d.part6}
\alias{psins.4d.part8} 
\alias{psins.6d}
\alias{psins.6d.part6}
\alias{psins.6d.part81}
\alias{psins.6d.part82} 
\alias{psins.6d.part83} 
\alias{psins.6d.part84}
\alias{scv.mat} 
\alias{ssq.decomp}
\alias{vec}
\alias{vech} 
\alias{vech.cat} 
\alias{which.mat}


\title{Internal functions in the ks library}
\description{
These functions are user-level but the user is not required to use them directly.   
}


\keyword{internal}
