\name{ks-internal}

\alias{kde.boundary}
\alias{Lpdiff}
\alias{matrix.sqrt}
\alias{mur}
\alias{nur}
\alias{nurs}
\alias{Qr}
\alias{Sdr}
\alias{Sdrv}
\alias{symconv.1d}
\alias{symconv.nd}

\title{Internal functions in the ks library}
\description{
These functions are user-level but which the user is not required to use directly.
}


\keyword{internal}
