\name{ks-internal}
\alias{getRow}
\alias{Lpdiff}
\alias{matrix.sqrt}
\alias{mur}
\alias{nur}
\alias{nurs}
\alias{Qr}
\alias{rowKpow}
\alias{Sdr}
\alias{Sdrv}
\alias{symconv.1d}
\alias{symconv.nd}
\alias{dwsupp}
\alias{reg.ucv}
\alias{kr}
\title{Internal functions in the ks library}

\description{
These functions are user-level but which the user is not required to use directly.
}

\value{The user is not required to use directly these outputs.}

\keyword{internal}
