\name{worldbank}
\docType{data}
\alias{worldbank}
\title{Development indicators from the World Bank Group}

\description{
This data set contains six development indicators for national
entities for the year 2011, which is the latest year for which they are
consistently available. 
}

\usage{data(worldbank)}

\format{A matrix with 7 columns and 218 rows.
  Each row corresponds to a country.
  The first column is the country,
  the second is the per capita carbon dioxide emissions (thousands Kg),
  the third is the per capita GDP (thousands of current USD),
  the fourth is the annual GDP growth rate (\%),
  the fifth is the annual inflation rate (\%),
  the sixth is the percentage of internet users in the population (\%),
  the seventh is the added value agricultural production as a ratio of the total GDP (\%). }

\source{
World Bank Group (2016) World development indicators. 
\url{http://databank.worldbank.org/data/reports.aspx?
source=world-development-indicators}. Accessed 2016-10-03.
}

\keyword{datasets}

